refactor segmentation workflows
move different types of workflows to seperate files. This will allow us to introduce optional segmentation dependencies for the different workflows in the future.
